+++ src/index.js
@@ -93,6 +93,10 @@ export default class Screen extends EventEmitter {
      * @param chars {String} the characters to print
      */
     write( chars ) {
+        if ( !chars ) {
+            return;
+        }
+
         // If there are no lines then we need to create one
         if ( this.cursor.x < 0 ) {
             this.createLine();
@@ -148,11 +152,46 @@ export default class Screen extends EventEmitter {
     }
 
     /**
+     * Writes a single character at the current cursor position
+     */
+    writechar( chars ) {
+        if ( !chars ) {
+            return;
+        }
+
+        // If there are no lines then we need to create one
+        if ( this.cursor.x < 0 ) {
+            this.createLine();
+            this.cursor.x = 0;
+            this.cursor.y = 0;
+        }
+
+        // If we're at the end of a line then we need to create a new one
+        if ( this.cursor.x >= this.opts.cols ) {
+            this.createLine();
+            this.cursor.x = 0;
+            this.cursor.y++;
+        }
+
+        // Now print the char
+        this.write( chars[ 0 ] );
+        this.emit( 'prompt', false );
+    }
+
+    /**
      * Writes the specified characters to a new line
      *
      * @param chars {String} the characters to print
      */
     writeln( chars ) {
+        if ( !chars ) {
+            this.createLine();
+            this.cursor.x = 0;
+            this.cursor.y = this.lines.length - 1;
+            this.emit( 'prompt', false );
+            return;
+        }
+
         // Split the input into separate lines to print
         var lines = this.splitLine( chars );
 
@@ -225,6 +264,13 @@ export default class Screen extends EventEmitter {
      */
     puts() {
         this.writeln.apply( this, arguments );
+    }
+
+    /**
+     * Putc alias - for writechar
+     */
+    putc() {
+        this.writechar.apply( this, arguments );
     }
 
     /**